apply shipping status to order-single, use custom line-items component
