+++ packages/shift-react-components/src/components/orders/order-list.js
@@ -4,6 +4,7 @@ import format from 'date-fns/format'
 import t from 'typy'
 
 // Lib
+import { fulfillmentStatus } from '../../lib/fulfillment-status'
 import { penceToPounds } from '../../lib/pence-to-pounds'
 
 export class OrderList extends PureComponent {
@@ -12,19 +13,7 @@ export class OrderList extends PureComponent {
 
     this.headers = props.headers || ['Order No.', 'Order Date', 'Cost', 'Shipping Status', '']
   }
-  // pending_approval unfulfilled short_shipped partially_shipped fully_shipped
 
-  fulfillmentStatus(status) {
-    const mappedStatuses = {
-      pending_approval: 'Awaiting shipment',
-      unfulfilled: 'Awaiting shipment',
-      short_shipped: 'Short shipped',
-      partially_shipped: 'Partially shipped',
-      fully_shipped: 'Shipped'
-    }
-
-    return mappedStatuses[status]
-  }
   /**
   * Render the order details
   * @param  {Object} order
@@ -53,7 +42,7 @@ export class OrderList extends PureComponent {
       },
       {
         label: this.headers[3],
-        value: this.fulfillmentStatus(order.fulfillment_status)
+        value: fulfillmentStatus(order.fulfillment_status)
       },
       {
         label: this.headers[4],
